Moved scheduler and synchronization code to "scheduler" lib
