+++ src/main/core/manager.rs
@@ -12,6 +12,7 @@ use atomic_refcell::AtomicRefCell;
 use log::warn;
 use rand::seq::SliceRandom;
 use rand_xoshiro::Xoshiro256PlusPlus;
+use scheduler::{HostIter, Scheduler, ThreadPerCoreSched, ThreadPerHostSched};
 use shadow_shim_helper_rs::emulated_time::EmulatedTime;
 use shadow_shim_helper_rs::shim_shmem::ManagerShmem;
 use shadow_shim_helper_rs::simulation_time::SimulationTime;
@@ -24,7 +25,6 @@ use crate::core::controller::{Controller, ShadowStatusBarState, SimController};
 use crate::core::cpu;
 use crate::core::resource_usage;
 use crate::core::runahead::Runahead;
-use crate::core::scheduler::{HostIter, Scheduler, ThreadPerCoreSched, ThreadPerHostSched};
 use crate::core::sim_config::{Bandwidth, HostInfo};
 use crate::core::sim_stats;
 use crate::core::worker;
