+++ src/main/core/mod.rs
@@ -5,7 +5,6 @@ pub mod logger;
 pub mod manager;
 pub mod resource_usage;
 pub mod runahead;
-pub mod scheduler;
 pub mod sim_config;
 pub mod sim_stats;
 pub mod work;
